Deprecate the Policy Path input variable
The PolicyPath variable has been removed from upstream
terraform-aws-lambda module, and we are no longer able to pass this
property into it.

The parameter's description has been updated to indicate that it's
deprecated, and it is no longer passed through to the
terraform-aws-lambda module from the main.tf file.

The use of the policy-path variable appears to have been removed several
versions earlier, within the terraform-aws-lambda module itself;
providing a value to this Kosli provider will have not had an effect.
The terraform-aws-lambda documentation never indicated that this was
deprecated or would be removed; it simply stopped working back in
v7.20.0 and was completely removed in v8.0.0